Updated solution to use size_t as opposed to int for consistency with std::string.

+++ cpp/hamming/hamming.cpp
@@ -1,11 +1,12 @@
 #include <string>
+#include <stdexcept>
 
 #include "hamming.h"
 
 
-unsigned int hamming::compute(const std::string& str1, const std::string& str2)
+size_t hamming::compute(const std::string& str1, const std::string& str2)
 {
-    unsigned int dist = 0;
+	size_t dist = 0;
     
     if (str1.length() != str2.length())
     {
@@ -14,7 +15,7 @@ unsigned int hamming::compute(const std::string& str1, const std::string& str2)
     else
     {
         // Iterate over the string and count differences
-        for (int i = 0; i < str1.length(); i++)
+        for (size_t i = 0; i < str1.length(); i++)
         {
             if (str1[i] != str2[i])
             {
